+++ VEngine/Code/Actors/IActorSystem.h
@@ -9,7 +9,7 @@ struct IActorSystem
 {
 	std::string name;
 
-	virtual void Tick(double deltaTime) = 0;
+	virtual void Tick(float deltaTime) = 0;
 	virtual void Init() = 0;
 	virtual std::vector<Actor*> GetActors() = 0;
 	virtual Actor* SpawnActor(Transform transform) = 0;
