+++ VEngine/Code/Actors/InteractTrigger.h
@@ -18,6 +18,6 @@ struct InteractTrigger : Actor
 
 	InteractTrigger();
 	virtual void Start() override;
-	virtual void Tick(double deltaTime) override;
+	virtual void Tick(float deltaTime) override;
 	virtual Properties GetProps() override;
 };
