+++ VEngine/Code/Components/Component.h
@@ -12,7 +12,7 @@ struct Component
 	std::string name;
 	int index = -1;
 
-	virtual void Tick(double deltaTime) = 0;
+	virtual void Tick(float deltaTime) = 0;
 	virtual void Start() {}
 	virtual void Create() = 0;
 	virtual void Destroy() = 0;
